pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Interview-learning RAG project using LangChain"
requires-python = ">=3.10"
dependencies = [
  "langchain",
  "langchain-community",
  "langchain-text-splitters",
  "langchain-chroma",
  "chromadb",
  "pypdf",
  "typer>=0.12",
  "python-dotenv>=1.0",
  "huggingface_hub>=0.22",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
]

[project.scripts]
rag = "rag_project.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
